util: Move utility functions for test infrastructure access
