+++ zenfs.mk
@@ -1,5 +1,18 @@
-zenfs_SOURCES = fs/fs_zenfs.cc fs/zbd_zenfs.cc fs/io_zenfs.cc
-zenfs_HEADERS = fs/fs_zenfs.h fs/zbd_zenfs.h fs/io_zenfs.h fs/version.h fs/metrics.h fs/snapshot.h
+zenfs_SOURCES = \
+	fs/fs_zenfs.cc \
+	fs/zbd_zenfs.cc \
+	fs/io_zenfs.cc \
+	fs/util.cc
+
+zenfs_HEADERS = \
+	fs/fs_zenfs.h \
+	fs/zbd_zenfs.h \
+	fs/io_zenfs.h \
+	fs/version.h \
+	fs/metrics.h \
+	fs/snapshot.h \
+	fs/util.h
+
 zenfs_LDFLAGS = -u zenfs_filesystem_reg
 
 ZENFS_ROOT_DIR := $(shell dirname $(realpath $(lastword $(MAKEFILE_LIST))))
